# app/backend/docs/swagger.yaml
definitions:
...
  dto.CreateParameterizationDto:
    properties:
      course_id:
        maximum: 100
        minimum: 1
        type: integer
      maxCreditsToOffer:
        maximum: 3
        minimum: 3
        type: integer
      numClassesPerDiscipline:
        maximum: 3
        minimum: 3
        type: integer
      semester_id:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - course_id
    - maxCreditsToOffer
    - numClassesPerDiscipline
    - semester_id
    type: object
...
  dto.UpdateParameterizationDto:
    properties:
      course_id:
        maximum: 100
        minimum: 1
        type: integer
      maxCreditsToOffer:
        maximum: 3
        minimum: 3
        type: integer
      numClassesPerDiscipline:
        maximum: 3
        minimum: 3
        type: integer
      semester_id:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - course_id
    - maxCreditsToOffer
    - numClassesPerDiscipline
    - semester_id
    type: object
...
  httperr.Fields:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
  httperr.RestErr:
    properties:
      code:
        type: integer
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/httperr.Fields'
        type: array
      message:
        type: string
    type: object
...
  response.ManyParameterizationsResponse:
    properties:
      parameterizations:
        items:
          $ref: '#/definitions/response.ParameterizationResponse'
        type: array
    type: object
...
  response.ParameterizationResponse:
    properties:
      course_id:
        type: integer
      maxCreditsToOffer:
        type: integer
      numClassesPerDiscipline:
        type: integer
      semester_id:
        type: integer
      uuid:
        type: string
    type: object
...
info:
  contact:
    email: your@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Time Table API
  version: "1.0"
paths:
...
  /parameterization/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete parameterization by uuid
      parameters:
      - description: parameterization uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Delete parameterization
      tags:
      - parameterization
  /parameterizations:
    post:
      consumes:
      - application/json
      description: Endpoint for create parameterization
      parameters:
      - description: Create parameterization dto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateParameterizationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Create new parameterization
      tags:
      - parameterization
  /parameterizations/{uuid}:
    get:
      consumes:
      - application/json
      description: Get parameterization by uuid
      parameters:
      - description: parameterization uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ParameterizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Parameterization details
      tags:
      - parameterization
    patch:
      consumes:
      - application/json
      description: Endpoint for update parameterization
      parameters:
      - description: parameterization uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: Update parameterization dto
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.UpdateParameterizationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Update parameterization
      tags:
      - parameterization
  /parameterizations/list-all:
    get:
      consumes:
      - application/json
      description: Get many parameterization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ManyParameterizationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Get many parameterization
      tags:
      - parameterization
  /parameterizations/list-all/{semesterId}:
    get:
      consumes:
      - application/json
      description: Get many parameterizations by semester
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ManyParameterizationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperr.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.RestErr'
      security:
      - ApiKeyAuth: []
      summary: Get many parameterizations by semester
      tags:
      - parameterizations by semester
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"